Create README.md
this is a project I have created in PowerBI, using the data in MsExcel ,.
The data analysis shows the graph of the various purchases in different cities and how the people reacted to different products. This gives the visualization of how to further take decisions and ultimately help the organization grow.